refactor: class LightController
