+++ app/CMakeLists.txt
@@ -11,7 +11,9 @@ add_executable(${PROJECT_NAME} ${sources} ${headers}
         src/MainController.cpp
         include/MainController.hpp
         src/MainEventController.cpp
-        include/MainEventController.hpp)
+        include/MainEventController.hpp
+        src/LightController.cpp
+        include/LightController.hpp)
 target_link_libraries(${PROJECT_NAME} PRIVATE matf-rg-engine)
 target_compile_features(${PROJECT_NAME} PRIVATE cxx_std_20)
 set_target_properties(${PROJECT_NAME} PROPERTIES RUNTIME_OUTPUT_DIRECTORY "${CMAKE_CURRENT_SOURCE_DIR}")
