+++ app/include/MainController.hpp
@@ -35,10 +35,6 @@ private:
 
     void draw_sphinx();
 
-    void draw_moon();
-
-    void draw_sun();
-
     void draw_camels();
 
     void draw() override;
@@ -48,24 +44,7 @@ private:
     void update_camera();
 
     bool m_cursor_enabled{true};
-    bool m_spotlight_enabled{false};
-    glm::vec3 m_sun_light{glm::vec3(255.0f, 255.0f, 224.0f) / 255.0f};
-    float m_sun_light_coeff{0.9f};
-    glm::vec3 m_moon_light{glm::vec3(247.0f, 234.0f, 198.0f) / 255.0f};
-    float m_moon_light_coeff{0.2f};
-    float m_ambient_coeff{0.1f};
 
-    struct PointLight {
-        glm::vec3 position;
-        glm::vec3 ambient;
-        glm::vec3 diffuse;
-        glm::vec3 specular;
-
-        float constant;
-        float linear;
-        float quadratic;
-
-    } m_point_light_sun, m_point_light_moon;
 };
 
 }// app
